Progress.

+++ src/main/kotlin/Nation.kt
@@ -1,7 +1,7 @@
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Nation(val name: String, val code: String, val owner: Int) {
+data class Nation(val name: String, val code: String, val owner: Long) {
     companion object {
         val NONE = Nation("none", "none", -1)
     }
